Add files via upload
Creates or defines the named selection of the faces of test parts to be used as sensor mounting surfaces in subsequent operations of this toolbox. If the named selection was not created before using this button, ensure that this button is re-run after the required named selection is generated and populated.