Update and rename plot_SG_calculations_FEA_v0.72.8.py to plot_SG_calculations_FEA_v0.72.9.py
"An SG_calculations.csv is found inside the solution directory. Would you like to plot the results for it instead?" now filters out comparison columns (ones that start with uppercase delta) if they exist in the input csv for the main data.

It now also filters out Time column as well from input csv file for the main data.